-add deleting files on exit

+++ HttpClient(.net 6.0)/Engine/FileManager.cs
@@ -5,7 +5,19 @@ namespace Engine;
 
 public class FileManager
 {
+    private static List<string> Caches = new List<string>();
 
+    public static void DeleteCaches()
+    {
+        foreach (var cache in Caches)
+        {
+            FileInfo file = new FileInfo(cache);
+            if (file.Exists)
+            {
+                file.Delete();
+            }
+        }
+    }
     public static String GetErorMessage(String status)
     {
         String message=$"<!DOCTYPE html> <html lang=\"ru\"><head><meta charset=\"utf-8\"><h1>Error</h1></head><h2>{status}</h2><body></body></html>";
@@ -16,13 +28,40 @@ public class FileManager
         File.AppendAllText("log.txt", message);
     }
 
-    public static void DeleteResponseFile(String filePath)
+    public static void DeleteResponseFile(String filePath, String url, bool finish = false)
     {
         FileInfo file = new FileInfo(filePath);
-        
-        if (file.Exists)
+        try
         {
-            file.Delete();
+            if (file.Exists)
+            {
+                if (finish == true)
+                {
+                    file.Delete();
+                    return;
+                }
+                if (CacheInfo.CacheTable.ContainsKey(url))
+                {
+                    var cacheResponse = CacheInfo.CacheTable[url];
+                    if (cacheResponse.cacheType != CacheType.NO_STORE)
+                    {
+                        if (!Caches.Contains(filePath))
+                        {
+                            Caches.Add(filePath);
+                        }
+
+                        return;
+                    }
+                    
+                    file.Delete();
+
+                    return; 
+                } 
+            }
+        }
+        catch (Exception e)
+        {
+            //Caches.Add(filePath);
         }
     }
